Moved image and surface data into the backend
There's no guarantee that these structures will be u32-indexes; in fact, on WebGL, they won't be
