+++ src/graphics/backend/surface_data.rs
@@ -1,0 +1,13 @@
+use graphics::backend::{Backend, BackendImpl};
+
+#[derive(Debug)]
+#[cfg(not(target_arch = "wasm32"))]
+pub struct SurfaceData {
+    pub framebuffer: u32
+}
+
+impl Drop for SurfaceData {
+    fn drop(&mut self) {
+        BackendImpl::destroy_surface(self);
+    }
+}
